# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/v1/tasks/": {
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Create task",
                "parameters": [
                    {
                        "description": "Create Params",
                        "name": "CreateParams",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_timohahaa_transcoder_internal_composer_modules_task.CreateForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task Model",
                        "schema": {
                            "$ref": "#/definitions/github_com_timohahaa_transcoder_internal_composer_modules_task.Task"
                        }
                    },
                    "default": {
                        "description": "Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_timohahaa_transcoder_internal_utils_render.HTTPError"
                        }
                    }
                }
            }
        },
        "/v1/tasks/{task_id}/": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Get task",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Task ID",
                        "name": "task_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task Model",
                        "schema": {
                            "$ref": "#/definitions/github_com_timohahaa_transcoder_internal_composer_modules_task.Task"
                        }
                    },
                    "default": {
                        "description": "Error",
                        "schema": {
                            "$ref": "#/definitions/render.HTTPError"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Delete task",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Task ID",
                        "name": "task_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response"
                    },
                    "default": {
                        "description": "Error",
                        "schema": {
                            "$ref": "#/definitions/render.HTTPError"
                        }
                    }
                }
            }
        },
        "/v1/tasks/{task_id}/cancel/": {
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Cancel task",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Task ID",
                        "name": "task_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response"
                    },
                    "default": {
                        "description": "Error",
                        "schema": {
                            "$ref": "#/definitions/render.HTTPError"
                        }
                    }
                }
            }
        },
        "/v1/tasks/{task_id}/progress/": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Get task progress",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Task ID",
                        "name": "task_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task progress",
                        "schema": {
                            "$ref": "#/definitions/tasks.taskProgress"
                        }
                    },
                    "default": {
                        "description": "Error",
                        "schema": {
                            "$ref": "#/definitions/render.HTTPError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_timohahaa_transcoder_internal_composer_modules_task.CreateForm": {
            "type": "object",
            "properties": {
                "duration": {
                    "type": "number"
                },
                "file_size": {
                    "type": "integer"
                },
                "settings": {
                    "$ref": "#/definitions/github_com_timohahaa_transcoder_internal_composer_modules_task.Settings"
                },
                "source": {
                    "$ref": "#/definitions/github_com_timohahaa_transcoder_internal_composer_modules_task.Source"
                }
            }
        },
        "github_com_timohahaa_transcoder_internal_composer_modules_task.Settings": {
            "type": "object",
            "properties": {
                "encrypt": {
                    "type": "boolean"
                }
            }
        },
        "github_com_timohahaa_transcoder_internal_composer_modules_task.Source": {
            "type": "object",
            "properties": {
                "fs": {
                    "$ref": "#/definitions/github_com_timohahaa_transcoder_internal_composer_modules_task.SourceFS"
                },
                "http": {
                    "$ref": "#/definitions/github_com_timohahaa_transcoder_internal_composer_modules_task.SourceHTTP"
                }
            }
        },
        "github_com_timohahaa_transcoder_internal_composer_modules_task.SourceFS": {
            "type": "object",
            "properties": {
                "path": {
                    "type": "string"
                }
            }
        },
        "github_com_timohahaa_transcoder_internal_composer_modules_task.SourceHTTP": {
            "type": "object",
            "properties": {
                "url": {
                    "type": "string"
                }
            }
        },
        "github_com_timohahaa_transcoder_internal_composer_modules_task.Task": {
            "type": "object",
            "properties": {
                "duration": {
                    "type": "number"
                },
                "encoder": {
                    "type": "string"
                },
                "file_size": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "routing": {
                    "type": "string"
                },
                "settings": {
                    "$ref": "#/definitions/github_com_timohahaa_transcoder_internal_composer_modules_task.Settings"
                },
                "source": {
                    "$ref": "#/definitions/github_com_timohahaa_transcoder_internal_composer_modules_task.Source"
                }
            }
        },
        "github_com_timohahaa_transcoder_internal_utils_render.HTTPError": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                },
                "invalid_params": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/validate.InvalidParamErr"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "render.HTTPError": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                },
                "invalid_params": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/validate.InvalidParamErr"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "tasks.taskProgress": {
            "type": "object",
            "properties": {
                "progress": {
                    "type": "integer"
                }
            }
        },
        "validate.InvalidParamErr": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "param_name": {
                    "type": "string"
                }
            }
        }
    }
}